Tiles scaled up 3x. Player speed reduced by 90%.

+++ Game/src/me/willchill/game/entity/Player.java
@@ -48,7 +48,7 @@ public class Player extends Entity{
 	}
 	
 	public void playerMove(double x, double y){
-		Game.level.xoffset -= (int) x;
-		Game.level.yoffset -= (int) y;
+		Game.level.xoffset -= x;
+		Game.level.yoffset -= y;
 	}
 }
